# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EventHouse.Management.Api",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://eventhouse-management-api-demo.onrender.com",
      "description": "Render Production"
    },
    {
      "url": "http://localhost:5185",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/v1/seatingMaps/{seatingMapId}": {
      "get": {
        "tags": [
          "SeatingMaps"
        ],
        "summary": "Retrieve a specific seating map by their unique identifier.",
        "operationId": "GetSeatingMapById",
        "parameters": [
          {
            "name": "seatingMapId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatingMapResponse"
                },
                "example": {
                  "id": "22222222-2222-2222-2222-222222222222",
                  "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                  "name": "Main Floor Seating",
                  "version": 1,
                  "isActive": true,
                  "createdAtUtc": "2026-06-20T21:40:59.5454705Z"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "put": {
        "tags": [
          "SeatingMaps"
        ],
        "summary": "Update an existing seating maps details.",
        "operationId": "UpdateSeatingMap",
        "parameters": [
          {
            "name": "seatingMapId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Idempotency header configured. Allowed methods: POST, PUT, DELETE",
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": "e5e6184b-f12b-405d-8926-2a4f7176fdc2"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSeatingMapRequest"
              },
              "example": {
                "version": 1,
                "name": "Main Floor Seating",
                "isActive": true
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSeatingMapRequest"
              },
              "example": {
                "version": 1,
                "name": "Main Floor Seating",
                "isActive": true
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSeatingMapRequest"
              },
              "example": {
                "version": 1,
                "name": "Main Floor Seating",
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "delete": {
        "tags": [
          "SeatingMaps"
        ],
        "summary": "Delete a seating map from the system.",
        "operationId": "DeleteSeatingMap",
        "parameters": [
          {
            "name": "seatingMapId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Idempotency header configured. Allowed methods: POST, PUT, DELETE",
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": "236b263a-62c7-456c-aafb-9e83951d8959"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/seatingMaps": {
      "get": {
        "tags": [
          "SeatingMaps"
        ],
        "summary": "List seating maps with optional filtering, sorting, and pagination.",
        "operationId": "ListSeatingMaps",
        "parameters": [
          {
            "name": "venueId",
            "in": "query",
            "description": "Venue ID to filter seating maps by.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Seating map name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "Indicates whether the seatingMap is currently active.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SeatingMapSortBy"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatingMapResponsePagedResult"
                },
                "example": {
                  "items": [
                    {
                      "id": "22222222-2222-2222-2222-222222222222",
                      "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                      "name": "Main Floor Seating",
                      "version": 1,
                      "isActive": true,
                      "createdAtUtc": "2026-06-20T21:40:59.5657303Z"
                    }
                  ],
                  "totalCount": 1,
                  "page": 1,
                  "pageSize": 20,
                  "links": {
                    "self": "/api/v1/seatingMaps?page=1\u0026pageSize=20",
                    "first": "/api/v1/seatingMaps?page=1\u0026pageSize=20",
                    "previous": null,
                    "next": null,
                    "last": "/api/v1/seatingMaps?page=1\u0026pageSize=20"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "post": {
        "tags": [
          "SeatingMaps"
        ],
        "summary": "Create a new seating map in the system.",
        "operationId": "CreateSeatingMap",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Idempotency header configured. Allowed methods: POST, PUT, DELETE",
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": "302f90c6-b490-488e-82f2-327de328a580"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSeatingMapRequest"
              },
              "example": {
                "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                "name": "Main Floor Seating",
                "isActive": true
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSeatingMapRequest"
              },
              "example": {
                "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                "name": "Main Floor Seating",
                "isActive": true
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CreateSeatingMapRequest"
              },
              "example": {
                "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                "name": "Main Floor Seating",
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "201": {
            "description": "Created",
            "headers": {
              "Location": {
                "description": "URL of the newly created resource",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatingMapResponse"
                },
                "example": {
                  "id": "22222222-2222-2222-2222-222222222222",
                  "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                  "name": "Main Floor Seating",
                  "version": 1,
                  "isActive": true,
                  "createdAtUtc": "2026-06-20T21:40:59.5742944Z"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSeatingMapRequest": {
        "required": [
          "name",
          "venueId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "type": "string",
            "description": "Seating map name."
          },
          "isActive": {
            "type": "boolean",
            "description": "Indicates whether the seatingMap is currently active."
          },
          "venueId": {
            "type": "string",
            "description": "The unique identifier of the venue to be associated with the seating map.",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "EventHouseProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "traceId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "PaginationLinks": {
        "type": "object",
        "properties": {
          "self": {
            "type": "string",
            "nullable": true
          },
          "first": {
            "type": "string",
            "nullable": true
          },
          "previous": {
            "type": "string",
            "nullable": true
          },
          "next": {
            "type": "string",
            "nullable": true
          },
          "last": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SeatingMapResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the seating map.",
            "format": "uuid"
          },
          "venueId": {
            "type": "string",
            "description": "Unique identifier of the venue.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Seating map name."
          },
          "version": {
            "type": "integer",
            "description": "Current version of the seating map (for concurrency control).",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indicates whether the seatingMap is currently active."
          },
          "createdAtUtc": {
            "type": "string",
            "description": "The date and time the seating map was created in UTC.",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "SeatingMapResponsePagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeatingMapResponse"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          }
        },
        "additionalProperties": false
      },
      "SeatingMapSortBy": {
        "enum": [
          "Name",
          "Version",
          "IsActive"
        ],
        "type": "string"
      },
      "SortDirection": {
        "enum": [
          "Asc",
          "Desc"
        ],
        "type": "string"
      },
...
      "UpdateSeatingMapRequest": {
        "required": [
          "name",
          "version"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "type": "string",
            "description": "Seating map name."
          },
          "isActive": {
            "type": "boolean",
            "description": "Indicates whether the seatingMap is currently active."
          },
          "version": {
            "type": "integer",
            "description": "The current version of the seating map. Must match the version on the server for the update to succeed.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "additionalProperties": {}
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "JWT Authorization header. Example: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:UNAUTHORIZED",
              "title": "Unauthorized",
              "status": 401,
              "detail": "A valid JWT access token is required.",
              "errorCode": "UNAUTHORIZED",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:FORBIDDEN",
              "title": "Forbidden",
              "status": 403,
              "detail": "You do not have permission to access this resource.",
              "errorCode": "FORBIDDEN",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:INTERNAL_ERROR",
              "title": "Internal Server Error",
              "status": 500,
              "detail": "An unexpected error occurred.",
              "errorCode": "INTERNAL_ERROR",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:NOT_FOUND",
              "title": "Not Found",
              "status": 404,
              "detail": "The requested resource was not found.",
              "errorCode": "NOT_FOUND",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:CONFLICT",
              "title": "Conflict",
              "status": 409,
              "detail": "The request conflicts with the current state of the resource.",
              "errorCode": "CONFLICT",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/ValidationProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:VALIDATION_ERROR",
              "title": "Validation error",
              "status": 400,
              "detail": "One or more validation errors occurred.",
              "errors": {
                "name": [
                  "The Name field is required."
                ]
              },
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Too Many Requests",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying the request.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:RATE_LIMIT_EXCEEDED",
              "title": "Too Many Requests",
              "status": 429,
              "detail": "Rate limit exceeded. Please retry later.",
              "errorCode": "RATE_LIMIT_EXCEEDED",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}